🚀[S08P22A201-148] 스팟 순위

See merge request s08-bigdata-dist-sub2/S08P22A201!25

+++ backend/src/main/java/com/a201/countingstar/db/entity/star/StarGrade.java
@@ -34,7 +34,7 @@ public class StarGrade {
     @Column(name="grade_1")
     private int grade1;
     @Column(name="grade_2")
-    private int grade2;
+    private double grade2;
 
     @Column(name="start_latitude")
     private String startLatitude;
